+++ test/unit/test_callback.py
@@ -3,7 +3,7 @@ import shutil
 import numpy as np
 import tensorflow as tf
 
-from deepreg.callback import CheckpointManagerCallback
+from deepreg.callback import build_callbacks, restore_model
 
 
 def test_restore_CheckpointManagerCallback():
@@ -29,40 +29,61 @@ def test_restore_CheckpointManagerCallback():
         return tf.data.Dataset.from_tensor_slices((inputs, labels)).repeat().batch(2)
 
     # train old_model and save
-    old_model = Net()
-    old_model.compile(optimizer=tf.keras.optimizers.Adam(0.1), loss=tf.keras.losses.MSE)
-    old_callback = CheckpointManagerCallback(
+    if len(tf.config.list_physical_devices("gpu")) > 1:
+        strategy = tf.distribute.MirroredStrategy()
+    else:  # use default strategy
+        strategy = tf.distribute.get_strategy()
+
+    with strategy.scope():
+        old_model = Net()
+        old_optimizer = tf.keras.optimizers.Adam(0.1)
+    old_model.compile(optimizer=old_optimizer, loss=tf.keras.losses.MSE)
+    old_callbacks = build_callbacks(
         model=old_model,
-        directory="./test/unit/ckpt_old",
-        period=5,
+        dataset=toy_dataset(),
+        log_dir="./test/unit/old",
+        histogram_freq=1,
+        save_period=5,
     )
     old_model.fit(
-        x=toy_dataset(), epochs=10, steps_per_epoch=10, callbacks=[old_callback]
+        x=toy_dataset(), epochs=10, steps_per_epoch=10, callbacks=old_callbacks
     )
 
     # create new model and restore old_model checkpoint
-    new_model = Net()
-    new_model.compile(optimizer=tf.keras.optimizers.Adam(0.1), loss=tf.keras.losses.MSE)
-    new_callback = CheckpointManagerCallback(
-        model=old_model, directory="./test/unit/ckpt_new"
+    with strategy.scope():
+        new_model = Net()
+        new_optimizer = tf.keras.optimizers.Adam(0.1)
+    new_model.compile(optimizer=new_optimizer, loss=tf.keras.losses.MSE)
+    new_callbacks = build_callbacks(
+        model=new_model,
+        dataset=toy_dataset(),
+        log_dir="./test/unit/new",
+        histogram_freq=1,
+        save_period=5,
     )
-    new_callback.restore(save_path="./test/unit/ckpt_old/ckpt-10")
+    initial_epoch = restore_model(
+        new_callbacks, ckpt_path="./test/unit/old/save/ckpt-10"
+    )
 
     # check equal
-    new_callback._manager.save()
-    old_reader = tf.train.load_checkpoint("./test/unit/ckpt_old/ckpt-10")
-    new_reader = tf.train.load_checkpoint("./test/unit/ckpt_new")
+    new_callbacks[1]._manager.save(0)
+    old_reader = tf.train.load_checkpoint("./test/unit/old/save/ckpt-10")
+    new_reader = tf.train.load_checkpoint("./test/unit/new/save")
     for k in old_reader.get_variable_to_shape_map().keys():
-        if "save_counter" not in k:
+        if "save_counter" not in k and "_CHECKPOINTABLE_OBJECT_GRAPH" not in k:
             equal = np.array(old_reader.get_tensor(k)) == np.array(
                 new_reader.get_tensor(k)
             )
-            assert equal.all(), k
+            assert equal, "{} fail to restore !".format(k)
+
+    new_model.fit(
+        x=toy_dataset(),
+        initial_epoch=initial_epoch,
+        epochs=20,
+        steps_per_epoch=10,
+        callbacks=new_callbacks,
+    )
 
     # remove temporary ckpt directories
-    shutil.rmtree("./test/unit/ckpt_old")
-    shutil.rmtree("./test/unit/ckpt_new")
-
-
-# if __name__ == '__main__':
-#     test_restore_CheckpointManagerCallback()
+    shutil.rmtree("./test/unit/old")
+    shutil.rmtree("./test/unit/new")