+++ CHANGELOG.md
@@ -17,6 +17,7 @@ adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Use CheckpointManager callback for saving and support training restore.
 - Changed distribute strategy to default for <= 1 GPU.
 - Migrated from Travis-CI to GitHub Actions.
 - Simplified configuration for backbone models.